Renamed private method (cherry picked from commit ad02ac3)

+++ beam-avhrr-reader/src/main/java/org/esa/beam/dataio/avhrr/noaa/CloudReader.java
@@ -57,7 +57,7 @@ class CloudReader implements BandReader {
         return ProductData.TYPE_UINT8;
     }
 
-    public float getScalingFactor() {
+    public double getScalingFactor() {
         return 1f;
     }
